OBJECT KIT: Make KVO actually work. This required a few fixes to the code for removing observers. I also genericised it along the way. There were other issues in this, which required...
RUNTIME: Add all the root zones, and don't rely on objects being
freed in a dependent order. This could be implemented by having
the tracer resolve all the dependencies it finds inside zones, but
I suspect that this could lead to memory never being freed. In the
case of circular references it certainly wouldn't work that simply.
Realistically, it would require at the least adding an extra flag
to the zone info structure. I will think about it another day.